+++ CHANGES.md
@@ -52,6 +52,8 @@ IronBee v0.9.0
 
 - Predicate evaluation state has been moved out of the Node subclasses and into a NodeEvalState class.  This improves const correctness and removes the dependence of Predicate on specific multithreading approaches.  In particular, Predicate now works with continuation approaches.
 
+- Utility functions like P.define(...) are moving to a new namespace (PUtil) and will all start with uppercase letters (e.g., PUtil.Define(...)).  Deprecation warnings are enabled and old naming conventions should be changed to the new format.
+
 **Fast**
 
 - Added `extract_waggle.rb` to extract fast patterns from waggle rules and updated `build.rb` to use appropriately for `.lua` and `.waggle` files.